stable -> master (#5901)
* search POC

* finishes ingesting the remainder of the database objects into the search table

* make fix and images

* migrtations reset

* feed img fix

* disables CLR round

* date

* chore: update sponsor clr section

* avatar url

* fixes the leaderboard job

* a few small tweaks

* enables daily email even iff there are no new bounties (#5831)

* change default trend settings

* sends new bounty daily email even if there are only actions (not new bounties

* one quick copy update

* Update emails.py

* stop clr to save some cpu

* is debugging offers https://gitcoincore.slack.com/archives/CAXQ7PT60/p1579891144044900

* kudos tab + toggle mode

* fixes syntax error; my bad

* loading indicator, new kdudos feed and corrects kudos feed kudos size

* fixes /home nav

* a fix to showing activity feed items that are hidden, also the ability to one click approve + email token requests

* email fixes

* prevent arbitrary created_on filter

* fixes derpy email subject line

* fixes https://sentry.io/organizations/gitcoin/issues/1448838094/?project=1398424&query=is%3Aunresolved&statsPeriod=14d&utc=true

* alert timing

* big scary warning to make sure u have access to addr u put int ot he form

* loading indicator parent actiivyt

* make fix; and a bailout for
https://github.com/gitcoinco/web/pull/5834

* offers are now 10x less lame looking

* fixes grant number

* fix offer backgrounds

* changes offers for the better by adding the ability to have an offerer name / link

* last little straggler commit

* default tab

* fixes a couple derpy things about actions

* fixes the daily action

* cancel bounty

* cleans up the /products page, significantly alts the nav for more consistency and takes users to the various marketplaces accoss the site as a default (an consistent about button is now on the nav tho so ppl can easily see what these produts are about if they care to read the docs)

*  usdt value was wrong

* only search if is staff

* onme small fix

* token creator

+++ scripts/debug/migrate_profile_picture.py
@@ -2,8 +2,9 @@ from app.utils import sync_profile
 from avatar.utils import get_avatar
 from dashboard.models import Profile
 
-handle='whalechat'
+handle='proofsuite'
 profile = Profile.objects.get(handle=handle)
 avatar = profile.avatar_baseavatar_related.first()
-avatar.delete()
+if avatar:
+    avatar.delete()
 sync_profile(profile.handle)
